+++ js/main.js
@@ -3,5 +3,6 @@ import './util.js';
 import './drawing-thumbnails.js';
 import './drawing-big-picture.js';
 import './form.js';
-import './img-preview.js';
+import './validate-form.js';
+import './scale-buttons.js';
 import './img-effects.js';
